requirements.txt
youtube-dl
awesome-slugify

